+++ src/libse/SubtitleFormats/AdvancedSubStationAlpha.cs
@@ -2518,7 +2518,7 @@ Format: Layer, Start, End, Style, Name, MarginL, MarginR, MarginV, Effect, Text"
                             }
                             else if (i == fontsizeIndex)
                             {
-                                if (float.TryParse(f, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture, out var fOut))
+                                if (decimal.TryParse(f, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture, out var fOut))
                                 {
                                     style.FontSize = fOut;
                                 }
